+++ src/frontend/apps/impress/src/features/docs/doc-export/blocks-mapping/index.ts
@@ -1,6 +1,7 @@
 export * from './dividerDocx';
 export * from './dividerPDF';
 export * from './headingPDF';
+export * from './imageDocx';
 export * from './imagePDF';
 export * from './paragraphPDF';
 export * from './quoteDocx';