+++ src/frontend/apps/impress/src/features/docs/doc-export/mappingDocx.tsx
@@ -2,8 +2,9 @@ import { docxDefaultSchemaMappings } from '@blocknote/xl-docx-exporter';
 
 import {
   blockMappingDividerDocx,
+  blockMappingImageDocx,
   blockMappingQuoteDocx,
-} from './blocks-mapping/';
+} from './blocks-mapping';
 import { DocsExporterDocx } from './types';
 
 export const docxDocsSchemaMappings: DocsExporterDocx['mappings'] = {
@@ -12,5 +13,6 @@ export const docxDocsSchemaMappings: DocsExporterDocx['mappings'] = {
     ...docxDefaultSchemaMappings.blockMapping,
     divider: blockMappingDividerDocx,
     quote: blockMappingQuoteDocx,
+    image: blockMappingImageDocx,
   },
 };